feat: place order

+++ app/src/main/java/com/selsela/takeefapp/data/order/model/order/Address.kt
@@ -15,9 +15,9 @@ data class Address(
     @SerializedName("is_fav")
     val isFav: Int = 0,
     @SerializedName("latitude")
-    val latitude: Int = 0,
+    val latitude: Double = 0.0,
     @SerializedName("longitude")
-    val longitude: Int = 0,
+    val longitude: Double = 0.0,
     @SerializedName("name")
     val name: String = "",
     @SerializedName("note")
